♻️ retarget package + imports for moved playback VMs
