+++ desktopApp/src/jvmMain/kotlin/com/calypsan/listenup/desktop/Main.kt
@@ -5,6 +5,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.application
 import androidx.compose.ui.window.rememberWindowState
 import com.calypsan.listenup.client.di.platformModule
+import com.calypsan.listenup.client.di.playbackPresentationModule
 import com.calypsan.listenup.client.di.sharedModules
 import com.calypsan.listenup.client.playback.AudioPlayer
 import com.calypsan.listenup.desktop.di.desktopAppModule
@@ -24,6 +25,7 @@ fun main() {
         modules(
             sharedModules + // From :shared module
                 platformModule + // From :composeApp desktopMain
+                playbackPresentationModule + // Shared playback VM bindings
                 desktopAppModule, // Desktop app specific
         )
     }
